ser-aluno-mock-mcp/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mcp>=0.9.0
httpx>=0.25.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
structlog>=23.2.0